+++ packages/frontend/src/components/MetricQueryData/UnderlyingDataModalContent.tsx
@@ -319,6 +319,7 @@ const UnderlyingDataModalContent: FC<Props> = () => {
             onlyRaw,
             showTableNames: true,
             columnOrder: [],
+            pivotColumns: undefined, // underlying data is always unpivoted
         });
         return csvResponse;
     };